# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ayo Indonesia Api",
        "description": "Ayo Indonesia Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/matches": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "get list of matches",
                "description": "Returns list of matches",
                "operationId": "getMatchesList",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Matches"
                ],
                "summary": "create new matches",
                "description": "create data matches",
                "operationId": "createMatches",
                "requestBody": {
                    "description": "create matches",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2022-07-28 15:00:00"
                                    },
                                    "home": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "away": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "player_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "type": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "team": {
                                                    "type": "text",
                                                    "example": "home"
                                                },
                                                "minute": {
                                                    "type": "number",
                                                    "example": "30"
                                                },
                                                "note": {
                                                    "type": "text",
                                                    "example": "free kick goals"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully Created"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/matches/{id}": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "get matches by id",
                "description": "get matches by id",
                "operationId": "getMatchesById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "get matches by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "update matches by id",
                "description": "update matches data",
                "operationId": "updateMatchesById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "update matches by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update matches by id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2022-07-28 15:00:00"
                                    },
                                    "home": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "away": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "player_id": {
                                                    "type": "number",
                                                    "example": ""
                                                },
                                                "type": {
                                                    "type": "number",
                                                    "example": ""
                                                },
                                                "team": {
                                                    "type": "text",
                                                    "example": ""
                                                },
                                                "minute": {
                                                    "type": "number",
                                                    "example": ""
                                                },
                                                "note": {
                                                    "type": "text",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "action": "create",
                                                "player_id": 1,
                                                "type": 1,
                                                "team": "home",
                                                "minute": 50
                                            },
                                            {
                                                "id": 11,
                                                "action": "update",
                                                "player_id": 1,
                                                "type": 1,
                                                "team": "home",
                                                "minute": 20
                                            },
                                            {
                                                "id": 10,
                                                "action": "delete",
                                                "player_id": 2,
                                                "type": 1,
                                                "team": "home",
                                                "minute": 30
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Matches"
                ],
                "summary": "delete matches by id",
                "description": "delete matches by Id",
                "operationId": "deleteMatchesById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete matches by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success - No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}